Make syncthing script generic, so I can use it elsewhere.
This changes the way syncthing reads the HTTPS key, so if syncthing is
already installed, then between the first run of the new 04-nginx and
the new 00-syncthing, the syncthing process will be unable to read the
key and will probably crashloop. This should resolve itself once the
new 00-syncthing has been run for the first time.